# core/vm: use uint256.Bytes32 and builtin copy to make MSTORE faster